requirements.txt
pygments
markdown2
Jinja2
gitpython
PyYAML